+++ exonum-java-binding/common/src/main/java/com/exonum/binding/common/proofs/list/CheckedListProofImpl.java
@@ -65,7 +65,7 @@ public class CheckedListProofImpl<E> implements CheckedListProof {
   }
 
   @Override
-  public HashCode getRootHash() {
+  public HashCode getIndexHash() {
     checkValid();
     return calculatedIndexHash;
   }